% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{add_frame}
\alias{add_frame}
\title{Add a frame for text diagrams}
\usage{
add_frame(input = "clipboard", output = "clipboard")
}
\arguments{
\item{input}{"clipboard" or the file path for the text diagram.}

\item{output}{"clipboard", file name, or cat.}
}
\value{
text diagram with a frame
}
\description{
Add a frame for text diagrams
}
