% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{deepdown}
\alias{deepdown}
\title{Convert a markdown or R markdown file into a markdeep file}
\usage{
deepdown(input, view = TRUE, remove_yaml = TRUE)
}
\arguments{
\item{input}{Character. The path of the (R) markdown file.}

\item{view}{Logical. Whether view the output in the web browser.}

\item{remove_yaml}{Logical. Whether remove the yaml header of the input file}
}
\value{
a html file
}
\description{
Convert a markdown or R markdown file into a markdeep file
}
\examples{
library(deepdown)
deepdown("mdr.Rmd")
input <- "d:/Seafile/pz/doing/rpkg/deepdown/inst/rmarkdown/templates/deepdown-book/skeleton/skeleton.Rmd"
}
