% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{deepdown_doc}
\alias{deepdown_doc}
\title{rosr pdf template}
\usage{
deepdown_doc(
  ...,
  template_name,
  keep_tex = TRUE,
  citation_package = "natbib",
  md_extensions = c("-autolink_bare_uris", "-auto_identifiers")
)
}
\arguments{
\item{..., }{keep_tex, citation_package, md_extensions
Arguments passed to \code{rmarkdown::pdf_document()}.}

\item{template_name}{the template name.}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{citation_package}{The LaTeX package to process citations, \code{natbib}
or \code{biblatex}. Use \code{default} if neither package is to be used,
which means citations will be processed via the command
\command{pandoc-citeproc}.}

\item{md_extensions}{Markdown extensions to be added or removed from the
default definition of R Markdown. See the \code{\link[rmarkdown]{rmarkdown_format}} for
additional details.}
}
\value{
An R Markdown output format.
}
\description{
rosr pdf template
}
\details{
statement_svm, article_svm, cv_svm, syllabus_svm, manuscript_svm are adapted from \url{https://github.com/svmiller/svm-r-markdown-templates}.
}
