% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{view_deepdown}
\alias{view_deepdown}
\title{View demo file}
\usage{
view_deepdown(template = "latex")
}
\arguments{
\item{template}{template name. see get_deepdown().}
}
\value{
a file in the viewer
}
\description{
View demo file
}
\examples{
view_deepdown()
}
